[Code]: fix new eslint errors

+++ x-pack/plugins/code/server/lsp/install_manager.test.ts
@@ -3,8 +3,7 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-
-/* tslint:disable */
+/* eslint-disable no-console */
 
 import fs from 'fs';
 import nock from 'nock';
@@ -14,12 +13,12 @@ import tar from 'tar-fs';
 import URL from 'url';
 import zlib from 'zlib';
 import { LanguageServers } from './language_servers';
-import { InstallManager } from "./install_manager";
-import { ServerOptions } from "../server_options";
+import { InstallManager } from './install_manager';
+import { ServerOptions } from '../server_options';
 import rimraf from 'rimraf';
 import { LanguageServerStatus } from '../../common/language_server';
 import { Server } from 'hapi';
-import { InstallationType } from "../../common/installation";
+import { InstallationType } from '../../common/installation';
 
 const LANG_SERVER_NAME = 'Java';
 const langSrvDef = LanguageServers.find(l => l.name === LANG_SERVER_NAME)!;
@@ -30,7 +29,7 @@ const fakeTarDir = path.join(fakeTestDir, 'fakePackage');
 const fakeFile = 'fake.file';
 const fakeContent = 'fake content';
 const options: ServerOptions = {
-  langServerPath: fakeTestDir
+  langServerPath: fakeTestDir,
 } as ServerOptions;
 
 const server = new Server();
@@ -43,9 +42,9 @@ server.config = () => {
     },
     has(key: string): boolean {
       return key === 'pkg.version';
-    }
-  }
-}
+    },
+  };
+};
 
 const manager = new InstallManager(server, options);
 
@@ -62,10 +61,11 @@ beforeAll(async () => {
       .pipe(fs.createWriteStream(fakePackageFile))
       .on('finish', resolve);
   });
-
 });
 beforeEach(() => {
-  const downloadUrl = URL.parse(langSrvDef.downloadUrl!(langSrvDef, server.config().get('pkg.version')));
+  const downloadUrl = URL.parse(
+    langSrvDef.downloadUrl!(langSrvDef, server.config().get('pkg.version'))
+  );
   nock.cleanAll();
   // mimic github's behavior, redirect to a s3 address
   nock(`${downloadUrl.protocol}//${downloadUrl.host!}`)
@@ -77,32 +77,29 @@ beforeEach(() => {
     .get('/file.tar.gz')
     .replyWithFile(200, fakePackageFile, {
       'Content-Length': fs.statSync(fakePackageFile).size.toString(),
-      'Content-Disposition': `attachment ;filename=${path.basename(fakePackageFile)}`
+      'Content-Disposition': `attachment ;filename=${path.basename(fakePackageFile)}`,
     });
 });
 afterEach(() => {
-  let p = manager.installationPath(langSrvDef);
-  if (p)
-    rimraf.sync(p);
+  const p = manager.installationPath(langSrvDef);
+  if (p) rimraf.sync(p);
 });
 afterAll(() => {
   nock.cleanAll();
   rimraf.sync(fakeTestDir);
 });
 
-
 test('it can download a package', async () => {
-  const manager = new InstallManager(server, options);
-  langSrvDef.installationType = InstallationType.Download
+  langSrvDef.installationType = InstallationType.Download;
   const p = await manager.downloadFile(langSrvDef);
   console.log('package downloaded at ' + p);
   expect(fs.existsSync(p)).toBeTruthy();
-  expect(fs.statSync(p).size).toBe(fs.statSync(fakePackageFile).size)
+  expect(fs.statSync(p).size).toBe(fs.statSync(fakePackageFile).size);
 });
 
 test('it can install language server', async () => {
   expect(manager.status(langSrvDef)).toBe(LanguageServerStatus.NOT_INSTALLED);
-  langSrvDef.installationType = InstallationType.Download
+  langSrvDef.installationType = InstallationType.Download;
   const installPromise = manager.install(langSrvDef);
   expect(manager.status(langSrvDef)).toBe(LanguageServerStatus.INSTALLING);
   await installPromise;
@@ -113,21 +110,20 @@ test('it can install language server', async () => {
   expect(fs.readFileSync(fakeFilePath, 'utf8')).toBe(fakeContent);
 });
 
-
 test('install language server by plugin', async () => {
-  langSrvDef.installationType = InstallationType.Plugin
+  langSrvDef.installationType = InstallationType.Plugin;
   expect(manager.status(langSrvDef)).toBe(LanguageServerStatus.NOT_INSTALLED);
   const testDir = path.join(fakeTestDir, 'test_plugin');
   fs.mkdirSync(testDir);
-  const pluginName = langSrvDef.installationPluginName as string
+  const pluginName = langSrvDef.installationPluginName as string;
   // @ts-ignore
   server.plugins = {
     [pluginName]: {
       install: {
-        path: testDir
-      }
-    }
+        path: testDir,
+      },
+    },
   };
   expect(manager.status(langSrvDef)).toBe(LanguageServerStatus.READY);
-  expect(manager.installationPath(langSrvDef)).toBe(testDir)
+  expect(manager.installationPath(langSrvDef)).toBe(testDir);
 });
